+++ modules/images/src/lib/parsers/parse-image.js
@@ -1,6 +1,6 @@
 import assert from '../utils/assert';
 import {isImageTypeSupported, getDefaultImageType} from '../category-api/image-type';
-import {getImageData} from '../category-api/parsed-image-api';
+import {getImageData, getImageDataAsync} from '../category-api/parsed-image-api';
 import parseToImage from './parse-to-image';
 import parseToImageBitmap from './parse-to-image-bitmap';
 import parseToNodeImage from './parse-to-node-image';
@@ -36,11 +36,60 @@ export default async function parseImage(arrayBuffer, options, context) {
   }
 
   // Browser: if options.image.type === 'data', we can now extract data from the loaded image
-  if (imageType === 'data') {
-    image = getImageData(image);
+  switch (imageType) {
+    case 'data':
+      // extract data and "delete" the no-longer needed image
+      const imageData = options.image._async
+        ? await getImageDataAsync(image)
+        : await getImageData(image);
+
+      // imagebitmaps can be released explicitly
+      if (image.close) {
+        image.close();
+      }
+
+      return imageData;
+
+    default:
+      return image;
   }
 
-  return image;
+  /*
+  let imagePromise;
+  switch (loadType) {
+    case 'imagebitmap':
+      image = await parseToImageBitmap(arrayBuffer, options, url);
+      break;
+    case 'image':
+      imagePromise = parseToImage(arrayBuffer, options, url);
+      break;
+    case 'data':
+      // Node.js loads imagedata directly
+      imagePromise = parseToNodeImage(arrayBuffer, options);
+      break;
+    default:
+      assert(false);
+  }
+
+  // Browser: if options.image.type === 'data', we can now extract data from the loaded image
+  switch (imageType) {
+    case 'data':
+      // extract data
+      return imagePromise
+        .then(image => getImageDataAsync(image, {includeImage: true}))
+        .then(imageData => {
+          // imageBitmap has a close method to dispose of graphical resources, call if available
+          if (imageData.image && imageData.image.close) {
+            imageData.image.close();
+          }
+          delete imageData.image;
+          return imageData;
+        });
+
+    default:
+      return imagePromise;
+  }
+  */
 }
 
 // Get a loadable image type from image type
@@ -53,7 +102,9 @@ function getLoadableImageType(type) {
       return getDefaultImageType();
     default:
       // Throw an error if not supported
-      isImageTypeSupported(type);
+      if (!isImageTypeSupported(type)) {
+        throw new Error(`@loaders.gl/images: image type ${type} not supported in this environment`);
+      }
       return type;
   }
 }