+++ modules/images/test/lib/category-api/parsed-image-api.spec.js
@@ -1,15 +1,16 @@
+/* global ImageBitmap */
 import test from 'tape-promise/tape';
 import {load} from '@loaders.gl/core';
 import {ImageLoader} from '@loaders.gl/images';
 
 // PARSED IMAGE API
 import {
-  getDefaultImageType,
   isImageTypeSupported,
   isImage,
   getImageType,
   getImageSize,
-  getImageData
+  getImageData,
+  getImageDataAsync
 } from '@loaders.gl/images';
 
 const IMAGE_TYPES = ['auto', 'image', 'imagebitmap', 'data'];
@@ -30,18 +31,11 @@ async function loadImages() {
 }
 
 test('Image Category#Parsed Image API imports', t => {
-  t.ok(getDefaultImageType, 'getDefaultImageType() is defined');
   t.ok(isImageTypeSupported, 'isImageTypeSupported() is defined');
   t.ok(isImage, 'isImage() is defined');
   t.ok(getImageType, 'getImageType() is defined');
   t.ok(getImageSize, 'getImageSize() is defined');
   t.ok(getImageData, 'getImageData() is defined');
-  t.end();
-});
-
-test('Image Category#getDefaultImageType', async t => {
-  const imageType = getDefaultImageType();
-  t.ok(IMAGE_TYPES.includes(imageType), 'Returns an expected image type');
   t.end();
 });
 
@@ -91,9 +85,26 @@ test('Image Category#getImageSize', async t => {
 test('Image Category#getImageData', async t => {
   const IMAGES = await loadImages();
   for (const image of IMAGES) {
-    t.equals(typeof getImageData(image), 'object', 'returns data');
+    const imageData = getImageData(image);
+    t.equals(typeof imageData, 'object', 'returns data');
+    // @ts-ignore
+    t.notOk(imageData.worker, 'processed on main-thread');
   }
   // @ts-ignore
   t.throws(() => getImageData('not an image'));
+  t.end();
+});
+
+test('Image Category#getImageDataAsync', async t => {
+  const IMAGES = await loadImages();
+  for (const image of IMAGES) {
+    const imageData = getImageData(image);
+    const imageData2 = await getImageDataAsync(image);
+    t.deepEquals(imageData.data, imageData2.data, 'returns correct data');
+    if (typeof ImageBitmap !== 'undefined' && image instanceof ImageBitmap) {
+      // @ts-ignore
+      t.ok(imageData2.worker, 'processed on worker');
+    }
+  }
   t.end();
 });